split qwen36 moe decode startup helpers
