+++ crates/runner/src/main.rs
@@ -25,6 +25,7 @@ mod qwen36_moe_mtp;
 mod qwen36_moe_mtp_loader;
 mod qwen36_moe_persistent_decode;
 mod qwen36_moe_prefetch;
+mod qwen36_moe_prompt;
 mod qwen36_moe_residency;
 mod qwen36_moe_session;
 mod qwen36_moe_speculative;
